#43 - 웹소켓 적용 후 패키지 구조 변경
