--- src/main/java/com/brogs/crm/domain/externalmessenger/SendMethod.java
@@ -1,9 +1,0 @@
-package com.brogs.crm.domain.externalmessenger;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-
-public enum SendMethod {
-    KAKAO, EMAIL, SMS, CONSOLE
-}
