+++ src/main/java/com/brogs/crm/domain/messagesystem/messenger/ExternalMessenger.java
@@ -1,7 +1,6 @@
-package com.brogs.crm.domain.externalmessenger;
+package com.brogs.crm.domain.messagesystem.messenger;
 
 import com.brogs.crm.domain.agentaccount.AccountCommand;
-import com.brogs.crm.domain.agentaccount.AccountInfo;
 
 public interface ExternalMessenger {
     void sendConfirmMessage(AccountCommand.RegisterProfile request, String confirmToken);
