+++ src/main/java/com/brogs/crm/domain/messagesystem/messenger/SendMethod.java
@@ -1,0 +1,6 @@
+package com.brogs.crm.domain.messagesystem.messenger;
+
+
+public enum SendMethod {
+    KAKAO, EMAIL, SMS, CONSOLE
+}
