+++ src/main/java/com/brogs/crm/infra/messengerApi/api/EmailMessageApiCaller.java
@@ -1,7 +1,7 @@
-package com.brogs.crm.infra.message.api;
+package com.brogs.crm.infra.messengerApi.api;
 
 import com.brogs.crm.domain.agentaccount.AccountCommand;
-import com.brogs.crm.domain.externalmessenger.SendMethod;
+import com.brogs.crm.domain.messagesystem.messenger.SendMethod;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
