+++ src/main/java/com/brogs/crm/infra/messengerApi/api/KakaoMessageApiCaller.java
@@ -1,7 +1,7 @@
-package com.brogs.crm.infra.message.api;
+package com.brogs.crm.infra.messengerApi.api;
 
 import com.brogs.crm.domain.agentaccount.AccountCommand;
-import com.brogs.crm.domain.externalmessenger.SendMethod;
+import com.brogs.crm.domain.messagesystem.messenger.SendMethod;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -10,16 +10,14 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ConsoleMessageApiCaller implements MessageApiCaller{
+public class KakaoMessageApiCaller implements MessageApiCaller{
     @Override
     public boolean support(SendMethod sendMethod) {
-        return SendMethod.CONSOLE == sendMethod;
+        return false;
     }
 
     @Override
     public void sendConfirmMail(AccountCommand.RegisterProfile request, String confirmToken) {
-        log.info("Console Message API Called !!");
-        log.info("인증코드 전송이 시작됩니다 console: [ {} ], code: [ {} }", request.getEmail(), confirmToken);
 
     }
 }
