+++ src/main/java/com/brogs/crm/infra/messengerApi/api/SmsMessageApiCaller.java
@@ -1,16 +1,15 @@
-package com.brogs.crm.infra.message.api;
+package com.brogs.crm.infra.messengerApi.api;
 
 import com.brogs.crm.domain.agentaccount.AccountCommand;
-import com.brogs.crm.domain.externalmessenger.SendMethod;
+import com.brogs.crm.domain.messagesystem.messenger.SendMethod;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
-
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class KakaoMessageApiCaller implements MessageApiCaller{
+public class SmsMessageApiCaller implements MessageApiCaller{
     @Override
     public boolean support(SendMethod sendMethod) {
         return false;
